Don't route on proposed ways
Routing on proposed ways was blacklisted only for car profile.
This adds the same blacklisting also for foot and bike profiles.

It looks like this behavior was added in
https://github.com/Project-OSRM/osrm-backend/pull/4258 and then
accidentally reverted when pull request with title "profiles api v2"
at https://github.com/Project-OSRM/osrm-backend/pull/4072 was merged.
This commit restores the original change for bike and foot profiles.